backend/tests/security/test_token_security.py
"""
Security tests for JWT token and cookie security
"""
import pytest
from jose import jwt
from tests.helpers.api_client import APIClient
from tests.helpers.generators import generate_email, generate_name, generate_strong_password, generate_project_name
from tests.helpers.auth import create_user_and_get_token, create_project


@pytest.mark.security
class TestJWTTokenSecurity:
    """Test JWT token security"""
    
    def test_tampered_token_signature_rejected(self, api_client: APIClient):
        """Token with tampered signature should be rejected"""
        # Create valid user
        email = generate_email()
        user_data, valid_token = create_user_and_get_token(
            api_client, email, generate_name(), generate_strong_password()
        )
        
        # Tamper with token by changing last character
        tampered_token = valid_token[:-5] + "XXXXX"
        
        api_client.set_token(tampered_token)
        response = api_client.get("/api/v1/users/me")
        
        assert response.status_code == 401
    
    def test_modified_token_payload_rejected(self, api_client: APIClient):
        """Token with modified payload should be rejected"""
        # Create valid token
        email = generate_email()
        user_data, valid_token = create_user_and_get_token(
            api_client, email, generate_name(), generate_strong_password()
        )
        
        # Try to decode and modify (will fail signature check)
        try:
            # This will fail because we don't have the secret key
            payload = jwt.decode(valid_token, options={"verify_signature": False})
            payload["sub"] = "00000000-0000-0000-0000-000000000000"
            
            # Try to use modified token (signature won't match)
            modified_token = jwt.encode(payload, "wrong_secret", algorithm="HS256")
            
            api_client.set_token(modified_token)
            response = api_client.get("/api/v1/users/me")
            
            assert response.status_code == 401
        except:
            # If jwt operations fail, that's also fine - token is rejected
            pass
    
    def test_missing_authorization_header(self, api_client: APIClient):
        """Request without Authorization header should be rejected"""
        response = api_client.get("/api/v1/users/me")
        
        assert response.status_code in [401, 403]
    
    def test_malformed_authorization_header(self, api_client: APIClient):
        """Malformed Authorization header should be rejected"""
        # Test various malformed headers
        malformed_tokens = [
            "InvalidToken",
            "sometoken",
        ]
        
        for token in malformed_tokens:
            api_client.set_token(token)
            response = api_client.get("/api/v1/users/me")
            # Should return 401 or 403
            assert response.status_code in [401, 403]
    
    def test_wrong_token_type_rejected(self, authenticated_client):
        """Using access token where refresh token expected (and vice versa)"""
        api_client, user_data, access_token = authenticated_client
        
        # This is implicitly tested by the refresh endpoint tests
        # Access token won't work for refresh endpoint
        # Refresh token won't work for regular endpoints
        
        # Just verify access token works for normal endpoints
        response = api_client.get("/api/v1/users/me")
        assert response.status_code == 200


@pytest.mark.security
class TestCookieSecurity:
    """Test cookie security properties"""
    
    def test_refresh_cookie_properties(self, api_client: APIClient):
        """Refresh token cookie should have secure properties"""
        # Create and verify user
        email = generate_email()
        password = generate_strong_password()
        
        api_client.post(
            "/api/v1/auth/signup",
            json={"email": email, "name": generate_name(), "password": password}
        )
        
        # Verify email
        verify_send = api_client.post(
            "/api/v1/auth/send-verification",
            json={"email": email}
        )
        import re
        code_match = re.search(r"CODE: (\d{6})", verify_send.json()["message"])
        code = code_match.group(1)
        
        api_client.post(
            "/api/v1/auth/verify-email",
            json={"email": email, "code": code}
        )
        
        # Sign in
        signin_response = api_client.post(
            "/api/v1/auth/signin",
            json={"email": email, "password": password}
        )
        
        # Check Set-Cookie header
        set_cookie = signin_response.headers.get("set-cookie", "")
        
        # Verify security flags
        assert "HttpOnly" in set_cookie
        assert "SameSite" in set_cookie
    
    def test_cookie_cleared_on_signout(self, api_client: APIClient):
        """Cookie should be cleared on signout"""
        # Create and signin user
        email = generate_email()
        password = generate_strong_password()
        
        api_client.post(
            "/api/v1/auth/signup",
            json={"email": email, "name": generate_name(), "password": password}
        )
        
        # Verify
        verify_send = api_client.post(
            "/api/v1/auth/send-verification",
            json={"email": email}
        )
        import re
        code_match = re.search(r"CODE: (\d{6})", verify_send.json()["message"])
        code = code_match.group(1)
        api_client.post(
            "/api/v1/auth/verify-email",
            json={"email": email, "code": code}
        )
        
        # Signin
        api_client.post(
            "/api/v1/auth/signin",
            json={"email": email, "password": password}
        )
        
        # Signout
        signout_response = api_client.post("/api/v1/auth/signout")
        assert signout_response.status_code == 200


@pytest.mark.security
class TestPasswordSecurity:
    """Test password security"""
    
    def test_password_not_returned_in_responses(self, api_client: APIClient):
        """Password should never be returned in any response"""
        email = generate_email()
        password = generate_strong_password()
        
        # Signup
        signup_response = api_client.post(
            "/api/v1/auth/signup",
            json={"email": email, "name": generate_name(), "password": password}
        )
        
        # Verify password not in response
        assert "password" not in str(signup_response.json())
        assert "hashed_password" not in str(signup_response.json())
        assert password not in str(signup_response.json())
    
    def test_password_strength_enforced(self, api_client: APIClient):
        """Weak passwords should be rejected"""
        weak_passwords = [
            "weak",              # Too short, no uppercase, no digit
            "weakpass",          # No uppercase, no digit
            "WEAKPASS",          # No lowercase, no digit
            "WeakPass",          # No digit
            "weak123",           # No uppercase
            "WEAK123",           # No lowercase
        ]
        
        for weak_pwd in weak_passwords:
            response = api_client.post(
                "/api/v1/auth/signup",
                json={
                    "email": generate_email(),
                    "name": generate_name(),
                    "password": weak_pwd
                }
            )
            # Should be rejected
            assert response.status_code in [400, 422]
    
    def test_password_hashed_in_database(self, api_client: APIClient):
        """Passwords should be stored hashed, not plaintext"""
        email = generate_email()
        password = generate_strong_password()
        
        # Create user
        signup_response = api_client.post(
            "/api/v1/auth/signup",
            json={"email": email, "name": generate_name(), "password": password}
        )
        assert signup_response.status_code == 201
        
        # Verify email
        verify_send = api_client.post(
            "/api/v1/auth/send-verification",
            json={"email": email}
        )
        import re
        code = re.search(r"CODE: (\d{6})", verify_send.json()["message"]).group(1)
        api_client.post(
            "/api/v1/auth/verify-email",
            json={"email": email, "code": code}
        )
        
        # Signin and get profile
        signin_response = api_client.post(
            "/api/v1/auth/signin",
            json={"email": email, "password": password}
        )
        access_token = signin_response.json()["accessToken"]
        
        api_client.set_token(access_token)
        profile_response = api_client.get("/api/v1/users/me")
        
        # Profile should not contain password in any form
        profile_data = str(profile_response.json())
        assert password not in profile_data
        assert "password" not in profile_response.json()


@pytest.mark.security
class TestAuthorizationBypassing:
    """Test attempts to bypass authorization"""
    
    def test_cannot_access_other_user_profile(self, two_authenticated_clients):
        """Cannot use token to access different user's resources"""
        (user1_data, token1), (user2_data, token2), api_client = two_authenticated_clients
        
        # User 1's token
        api_client.set_token(token1)
        profile = api_client.get("/api/v1/users/me")
        
        # Should get User 1's profile, not User 2's
        assert profile.json()["id"] == user1_data["id"]
        assert profile.json()["email"] == user1_data["email"]
    
    def test_cannot_modify_other_user_resources(self, two_authenticated_clients):
        """Cannot modify resources owned by other users"""
        (user1_data, token1), (user2_data, token2), api_client = two_authenticated_clients
        
        # User 1 creates project
        api_client.set_token(token1)
        project = create_project(api_client, generate_project_name())
        
        # User 2 tries to create review in User 1's project
        api_client.set_token(token2)
        response = api_client.post(
            "/api/v1/reviews/",
            json={"name": "Unauthorized Review", "project_id": project["id"]}
        )
        
        # Should be forbidden
        assert response.status_code in [403, 500]
    
    def test_uuid_enumeration_protection(self, authenticated_client):
        """Non-existent UUIDs should return 404, not expose system info"""
        api_client, user_data, access_token = authenticated_client
        
        # Try to access non-existent resources
        fake_uuids = [
            "00000000-0000-0000-0000-000000000000",
            "11111111-1111-1111-1111-111111111111",
            "ffffffff-ffff-ffff-ffff-ffffffffffff",
        ]
        
        for fake_id in fake_uuids:
            # Try to complete non-existent checklist
            response = api_client.put(f"/api/v1/checklists/{fake_id}/complete")
            # Should return 404, not expose internal errors
            assert response.status_code in [404, 401, 403]

